Rename theme module to lowercase
